+++ pycricinfo/api/endpoints/seasons.py
@@ -2,6 +2,7 @@ from typing import Literal
 
 from fastapi import APIRouter, Path, Query, status
 
+from pycricinfo.config import MatchTypeNames
 from pycricinfo.search.seasons import get_match_types_in_season
 from pycricinfo.search.series import extract_match_ids_from_series
 from pycricinfo.source_models.pages.series import MatchResult, MatchType
@@ -19,26 +20,11 @@ router = APIRouter(prefix="/matches", tags=["matches"])
     summary="Get a list of match types, each containing a list of series in that match type for this season",
 )
 async def match_types_in_season(
-    season_name: int = Path(description='The name of the season to get matches for, e.g. "2024" or "2020-21"'),
-    match_type_name: Literal[
-        "Tests",
-        "One-Day Internationals",
-        "Twenty20 Internationals",
-        "First-class",
-        "List A",
-        "Twenty20",
-        "Women's One-Day Internationals",
-        "Women's Twenty20 Internationals",
-        "Youth Tests",
-        "Youth One-Day Internationals",
-        "Women's Twenty20",
-        "Tour",
-        "Minor tour",
-        "Other matches",
-        "Other one-day/limited-overs matches",
-        "Other Twenty20 matches",
-    ] = Query(default=None, description="Filter the response to just matches of the named type"),
+    season_name: int|str = Path(description='The name of the season to get matches for, e.g. "2024" or "2020-21"'),
+    match_type_name: Literal[tuple(c.value for c in MatchTypeNames)] = Query(   # type: ignore
+        default=None, description="Filter the response to just matches of the named type"),
 ) -> list[MatchType]:
+    season_name = season_name.replace("-", "/") if isinstance(season_name, str) else season_name
     match_types = get_match_types_in_season(season_name)
 
     if match_type_name: