+++ tests/test_loading.py
@@ -12,7 +12,7 @@ def get_test_file_path(file_name: str) -> str:
     return os.path.join(os.path.dirname(__file__), "test_files", file_name)
 
 
-@pytest.mark.parametrize("file_name", ["match/1426555.json", "match/1243385.json"])
+@pytest.mark.parametrize("file_name", ["match/1426555.json", "match/1243385.json", "match/1381212.json"])
 def test_load_match(file_name):
     """Test loading JSON files into Match model without exceptions."""
     test_file_path = get_test_file_path(file_name)